adding info on factoryCalData.h
factoryCalData.h is used by program to store calibration data into flash of arduino for faster reading than from thermopile and allows for most data to be calculated almost instantly than stored in ram. in fact the slowest process of this sensor reading besides the math is the i2c communication.