make code compile without throwing minor errors. it still would compile it just didn't like the fast q_sqrt routine because it was not standard
removed the Q_sqrt routine from use. i believe it is faster, but it does make the compiler complain (it still will compile with it)
also removed the limitations on the math.h file it is now included no matter if NEW_METHOD true ;or false.
the compiler will decide if any parts of the math.h file are needed. if q_sqrt function  is un-remarked out and sqrt is replaced by it complier will have no need of math.h and this will result in ~5k less of flash storage. of course this depends on the math library and if it is setup for 8bit. 16 bit or 32bit flash memory, ect....,