made it so errors are thrown less often. (it still compiled)
added
   int occRow=0;
    int occColumn=0;
   int accRow=0;
    int accColumn=0;

this just defined and sets the start value at 0, the compiler is not smart enough to see that these values are always set from the pixel number. it has little effect on performance, and only uses a few more bytes in flash rom. all we are doing is initializing before we use the values. this prevents an error that states occ can be used without being defined.