changes to double resolution mode, added failsafes to freq that can be used for i2c on sensor
if arduino 80000khz
if teensy 1mhz
code will not compile unless sensor is within spec. prevents damage

double resolution is now a simple low memory  function

added in a ram cache that can be used if spare ram, however there are some issues to still resolve on the i2c and serial busses.
this ram cache is about 1.5 k currently. when issues resolved it will go down to 128bytes. default is disabled.