fixed the  error and noise at bottom of doubleread
the code is setup to cache on even y numbers from the upsample, the farthest left it can not interpolate, and the bottom it can not interpolate. for some reason i had assumed it was the top, but there are only 24 lines to read from so it os 0 to 23. because of the direction of the cache the last line 47 needs to be the same as 46. the cache updates at 46, and 47 uses that cache data.